reorganized tests folders
